requirements/base.txt
Django==5.0.8
psycopg2-binary==2.9.10
python-dotenv==1.0.1
whitenoise==6.2.0
django-extensions==3.2.1
pillow==10.4.0
django-ckeditor-5==0.2.13
django_widget_tweaks==1.5.0
jdatetime==5.0.0
django-solo==2.3.0
django-debug-toolbar==4.4.6